exception test solved
